+++ serialization/src/JSONSerializer.hpp
@@ -85,14 +85,14 @@ namespace sprawl
 			friend class ReplicableBase;
 			using SerializerBase::serialize;
 			template<typename T>
-			void serialize_impl(T* var, const size_t bytes, const std::string& name, bool)
+			void serialize_impl(T* var, const uint32_t bytes, const std::string& name, bool)
 			{
 				if(!m_bIsValid)
 				{
 					throw ex_invalid_data();
 				}
 				bool bIsArray = false;
-				size_t size = bytes/sizeof(T);
+				uint32_t size = bytes/sizeof(T);
 				if(bytes > sizeof(T))
 				{
 					StartArray(name, size, true);
@@ -174,7 +174,7 @@ namespace sprawl
 					std::stringstream converter;
 					if(bIsArray)
 					{
-						for(size_t i = 0; i < size; i++)
+						for(uint32_t i = 0; i < size; i++)
 						{
 							converter.str("");
 							converter << var[i];
@@ -207,7 +207,7 @@ namespace sprawl
 				}
 			}
 
-			void serialize_impl(char* var, const size_t, const std::string& name, bool)
+			void serialize_impl(char* var, const uint32_t, const std::string& name, bool)
 			{
 				if(!m_bIsValid)
 				{
@@ -286,7 +286,7 @@ namespace sprawl
 				}
 			}
 
-			void serialize_impl(std::string* var, const size_t, const std::string& name, bool)
+			void serialize_impl(std::string* var, const uint32_t, const std::string& name, bool)
 			{
 				if(!m_bIsValid)
 				{
@@ -365,81 +365,81 @@ namespace sprawl
 				}
 			}
 
-			virtual void serialize(int* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(int* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(long int* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(long int* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(long long int* var, const size_t bytes, const std::string& name, bool PersistToDB)  override
+			virtual void serialize(long long int* var, const uint32_t bytes, const std::string& name, bool PersistToDB)  override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(short int* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(short int* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(char* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(char* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(float* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(float* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(double* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(double* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(long double* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(long double* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(bool* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(bool* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(unsigned int* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(unsigned int* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(unsigned long int* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(unsigned long int* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(unsigned long long int* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(unsigned long long int* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(unsigned short int* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(unsigned short int* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual void serialize(unsigned char* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(unsigned char* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
-			virtual void serialize(std::string* var, const size_t bytes, const std::string& name, bool PersistToDB) override
+			virtual void serialize(std::string* var, const uint32_t bytes, const std::string& name, bool PersistToDB) override
 			{
 				serialize_impl(var, bytes, name, PersistToDB);
 			}
 
-			virtual size_t StartObject(const std::string& str, bool = true) override
+			virtual uint32_t StartObject(const std::string& str, bool = true) override
 			{
 				State LastState = m_stateTracker.empty() ? State::None : m_stateTracker.back();
 				m_stateTracker.push_back(State::Object);
@@ -535,7 +535,7 @@ namespace sprawl
 				}
 			}
 
-			virtual void StartArray(const std::string& str, size_t& size, bool = true) override
+			virtual void StartArray(const std::string& str, uint32_t& size, bool = true) override
 			{
 				State LastState = m_stateTracker.empty() ? State::None : m_stateTracker.back();
 				m_stateTracker.push_back(State::Array);
